0.4.0

+++ src/main/scala/licos/json/element/lobby/client2server/JsonIdSearch.scala
@@ -4,13 +4,13 @@ import licos.json.element.lobby.TypeSystem
 import licos.json.validation.lobby.{IdSearchValidation, LobbyValidation}
 import licos.json.validation.village.AvatarValidation
 
-final case class JsonIdSearch(`type`: String, token: String, lobby: String, idForSearching: Int)
+final case class JsonIdSearch(`type`: String, token: String, lobby: String, idForSearching: Long)
     extends TypeSystem(`type`) {
 
   override protected def validType: String = JsonIdSearch.`type`
 
   @SuppressWarnings(Array[String]("org.wartremover.warts.Overloading"))
-  def this(token: String, lobby: String, idForSearching: Int) = {
+  def this(token: String, lobby: String, idForSearching: Long) = {
     this(JsonIdSearch.`type`, token, lobby, idForSearching)
   }
 }
@@ -28,7 +28,7 @@ object JsonIdSearch {
     (JsPath \ "type").read[String](pattern(`type`.r)) and
       (JsPath \ "token").read[String](AvatarValidation.token) and
       (JsPath \ "lobby").read[String](LobbyValidation.lobby) and
-      (JsPath \ "idForSearching").read[Int](IdSearchValidation.idForSearching)
+      (JsPath \ "idForSearching").read[Long](IdSearchValidation.idForSearching)
   )(JsonIdSearch.apply _)
 
   implicit val jsonWrites: OWrites[JsonIdSearch] = Json.writes[JsonIdSearch]
